+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Accounts.java
@@ -1,10 +1,13 @@
 package com.techelevator.tenmo.model;
 
+import java.math.BigDecimal;
+
 public class Accounts {
 	
 	private long accountId;
 	private long userId;
-	private double balance;
+//	private double balance;
+	private BigDecimal balance;
 	
 	
 	public Accounts() {
@@ -23,10 +26,10 @@ public class Accounts {
 	public void setUserId(long userId) {
 		this.userId = userId;
 	}
-	public double getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
-	public void setBalance(double balance) {
+	public void setBalance(BigDecimal balance) {
 		this.balance = balance;
 	}
 	@Override
